requirements.txt
# Windows Service
pywin32>=311

# SSE Client
sseclient-py>=1.8.0

# HTTP Client
requests>=2.31.0

# System utilities
psutil>=5.9.0

# Configuration
PyYAML>=6.0

# Database (legacy SQL support)
pymssql>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# CSV generation
csvkit>=1.12.0

# MinIO Storage (Pattern C)
minio>=7.2.0

# WebSocket Client (Transport Layer)
websocket-client>=1.6.0
